Blade permissions

See merge request mmkrana/enswadesh!52

+++ app/Http/Controllers/Backend/Location/AreaController.php
@@ -9,7 +9,6 @@ use Illuminate\Http\Request;
 use App\Models\Location\Area;
 use App\Models\Location\City;
 use App\Http\Controllers\Controller;
-use Intervention\Image\ImageManager;
 
 class AreaController extends Controller
 {
@@ -19,6 +18,9 @@ class AreaController extends Controller
      */
     public function index()
     {
+        //Define Area authorize gate
+        Gate::authorize('backend.areas.index');
+
         $areas = Area::with('cityOfArea')->get();
         return view('backend.location.area.index',  compact('areas'));
     }
@@ -29,6 +31,9 @@ class AreaController extends Controller
      */
     public function create()
     {
+        //Define Area authorize gate
+        Gate::authorize('backend.areas.create');
+
         $cities = City::all();
         return view('backend.location.area.form', compact('cities'));
     }
@@ -80,6 +85,9 @@ class AreaController extends Controller
      */
     public function edit($id)
     {
+        //Define Area authorize gate
+        Gate::authorize('backend.areas.edit');
+
         $cities = City::all();
         $area = area::find($id);
         return view('backend.location.area.form', compact('area', 'cities'));
@@ -128,6 +136,9 @@ class AreaController extends Controller
      */
     public function destroy($id)
     {
+        //Define Area authorize gate
+        Gate::authorize('backend.areas.destroy');
+
         $data = Area::find($id);
         $oldFilename = $data->area_icon;
         Storage::delete('/uploads/shopproperty/area/' . $oldFilename);
